fix: merge 과정에서 패키지 구조 달라 생긴 오류 해결
