+++ src/main/java/org/choon/careerbee/domain/competition/repository/CompetitionResultRepository.java
@@ -1,7 +1,7 @@
 package org.choon.careerbee.domain.competition.repository;
 
 import org.choon.careerbee.domain.competition.domain.CompetitionResult;
-import org.choon.careerbee.domain.competition.repository.custom.CompetitionResultCustomRepository;
+import org.choon.careerbee.domain.competition.repository.custom.result.CompetitionResultCustomRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CompetitionResultRepository extends
