--- src/main/java/org/choon/careerbee/domain/competition/repository/custom/CompetitionResultCustomRepositoryImpl.java
@@ -1,96 +1,0 @@
-package org.choon.careerbee.domain.competition.repository.custom;
-
-import static org.choon.careerbee.domain.competition.domain.QCompetitionResult.competitionResult;
-
-import com.querydsl.core.Tuple;
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import org.choon.careerbee.domain.competition.dto.response.LiveRankingResp;
-import org.choon.careerbee.domain.competition.dto.response.LiveRankingResp.RankerInfo;
-import org.choon.careerbee.domain.competition.dto.response.MemberLiveRankingResp;
-import org.springframework.stereotype.Repository;
-
-@RequiredArgsConstructor
-@Repository
-public class CompetitionResultCustomRepositoryImpl implements CompetitionResultCustomRepository {
-
-    private final JPAQueryFactory queryFactory;
-
-    @Override
-    public Optional<MemberLiveRankingResp> fetchMemberLiveRankingByDate(
-        Long accessMemberId, LocalDate today
-    ) {
-        List<Tuple> todayCompetitionResults = queryFactory
-            .select(
-                competitionResult.member.id,
-                competitionResult.solvedCount,
-                competitionResult.elapsedTime
-            )
-            .from(competitionResult)
-            .where(competitionResult.createdAt.year().eq(today.getYear())
-                .and(competitionResult.createdAt.month().eq(today.getMonthValue()))
-                .and(competitionResult.createdAt.dayOfMonth().eq(today.getDayOfMonth())))
-            .orderBy(
-                competitionResult.solvedCount.desc(),
-                competitionResult.elapsedTime.asc())
-            .fetch();
-
-        Long rank = 1L;
-        for (Tuple resp : todayCompetitionResults) {
-            if (resp.get(competitionResult.member.id).equals(accessMemberId)) {
-                return Optional.of(new MemberLiveRankingResp(
-                    rank,
-                    resp.get(competitionResult.elapsedTime),
-                    resp.get(competitionResult.solvedCount)
-                ));
-            }
-            rank++;
-        }
-
-        return Optional.empty();
-    }
-
-    @Override
-    public LiveRankingResp fetchLiveRankingByDate(LocalDate today) {
-        List<Tuple> top10 = queryFactory
-            .select(
-                competitionResult.member.id,
-                competitionResult.member.nickname,
-                competitionResult.member.imgUrl, // Todo : 추후 badge url도 도입해야함
-                competitionResult.solvedCount,
-                competitionResult.elapsedTime
-            )
-            .from(competitionResult)
-            .where(competitionResult.createdAt.year().eq(today.getYear())
-                .and(competitionResult.createdAt.month().eq(today.getMonthValue()))
-                .and(competitionResult.createdAt.dayOfMonth().eq(today.getDayOfMonth())))
-            .orderBy(
-                competitionResult.solvedCount.desc(),
-                competitionResult.elapsedTime.asc())
-            .limit(10)
-            .fetch();
-
-        List<RankerInfo> rankerInfos = new ArrayList<>();
-
-        Long rank = 1L;
-        for (Tuple tuple : top10) {
-            rankerInfos.add(
-                new RankerInfo(
-                    rank,
-                    tuple.get(competitionResult.member.nickname),
-                    tuple.get(competitionResult.member.imgUrl),
-                    tuple.get(competitionResult.member.imgUrl), // Todo : 이후 badge url로 변경해야함
-                    tuple.get(competitionResult.elapsedTime),
-                    tuple.get(competitionResult.solvedCount)
-                )
-            );
-            rank++;
-        }
-
-        return new LiveRankingResp(rankerInfos);
-    }
-}
